COMMERCE-2475 dispatch - extract manual class from autogenerated ones
